[#1] allow passing full module paths instead of relying on automagic
Can pass in a full absolute path to `module` now.

`_module_dir` is now `_modules_dirs`, an array of possible PKCS#11 module
locations.

If passing in just a module name to `module`, all `_module_dirs` will be checked
to see if the file exists.

Using Types::Tiny and Types::Path::Tiny for some `isa` checks and so not using
Sub::Quote in isa checks anymore.